added ethernet_status_hist

+++ src/base/utils/hist.cc
@@ -48,7 +48,7 @@ void Hist::Push(int32_t value)
 }
 
 //used for robot_in_range_detected
-void Hist::Push2(int32_t value)
+void Hist::Push2(uint8_t value)
 {
         uint32_t temp_sum=0;
         for(int i=0;i<8;i++)
@@ -56,7 +56,7 @@ void Hist::Push2(int32_t value)
                 temp_sum = (sum2 >> (i*4)) & 0xF;
                 temp_sum = temp_sum - (((valueHist[valPointer] & (1<<i)) ==0 ) ? 0 : 1) + (((value & (1<<i))==0) ? 0 : 1);
                 sum2 &= ~(0xF << (i*4));
-                sum2 |= (temp_sum & 0xF) << (i*4);//MAX 15
+                sum2 |= (temp_sum & 0xF) << (i*4);//MAX 15 
         }
         valueHist[valPointer] = value;
         valPointer++;
